src/main/java/com/example/sobesresttask/service/impl/UserServiceImpl.java
package com.example.sobesresttask.service.impl;

import com.example.sobesresttask.model.User;
import com.example.sobesresttask.repository.UserRepository;
import com.example.sobesresttask.service.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class UserServiceImpl implements UserService {
    @Autowired
    private UserRepository userRepository;

    public User createUser(User user) {
        return userRepository.save(user);
    }

    @Override
    public Optional<User> getUserById(Long id) {
        return userRepository.findById(id);
    }
}
